Reworked the controls to meet my own gut. My first instinct kept screwing me up with the different things and the "anyUp" for the set launch game thing was a nightmare and would pop up all the time when I was looking through my games!
New Bindings:
Minus = Select Launch Game
Home = Opens and Closes the "QuickMenu" which is no longer a quick menu without the old circular design it is so I understand the old binding but it's more of a main menu now so the triggers felt bizarre to use to access it.
Dpad Up/Down and L/R = all switch between homebrew and base games. I may tweak this down to one or the other but "anyUp" or down would be just as annoying but I do feel like using the dpad to "scroll" through libraries might be nice
Plus = Still Opens Credits as this seems fair